* transfering predictions as int64 * using last_noise to reverse the previous noise images, instead of always going full std::fill (excet first frame) * rewritten draw code for FrameGrabber with tags in mind * apparently Tasks were deleted before they were applied in RecTasks. rewritten some parts that will hopefully compile everywhere :(

+++ Application/pretrained_tagwork.py
@@ -54,7 +54,7 @@ def predict():
         print("# empty images array")
     else:
         #print("# predicting ", len(tag_images))
-        receive(tagwork.predict(tag_images))
+        receive(tagwork.predict(tag_images).astype(np.int64))
 
         #del tag_images
